hw3/requirements.txt
functions-framework==3.2.1
Flask==2.2.5
google-cloud-storage==2.11.0
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0